Skrivit testkod, fått ett grönt och ett rött ljus i middleware-testkoden

+++ server/controllers/noteController.js
@@ -9,7 +9,7 @@ export const noteController = {
    * @param {*} res 
    * @param {*} next 
    */
-  create: async (err, req, res, next) => {
+  create: async (req, res, next) => {
     try {
       const savedNote = await new Note(req.body).save()
       res.status(201).json(savedNote)
@@ -26,7 +26,7 @@ export const noteController = {
    * @param {*} res 
    * @param {*} next 
    */
-  get: async (err, req, res, next) => {
+  get: async (req, res, next) => {
     try {
       const notes = await Note.find().populate('owner', 'email')
       res.json(notes)
@@ -43,13 +43,15 @@ export const noteController = {
    * @param {*} res 
    * @param {*} next 
    */
-  edit: async (err, req, res, next) => {
+  edit: async (req, res, next) => {
     try {
       const { note } = res.locals
+      if (!note) return res.status(404).json({ message: "Note not found" })
+        
       Object.assign(note, req.body)
 
       await note.save()
-      res.status(204).json(`Note ${note.title} is updated successfully`)
+      res.status(200).json(`Note ${note.title} is updated successfully`)
     } catch (err) {
       res.status(400).json({ error: err.message })
     }
@@ -64,8 +66,10 @@ export const noteController = {
    * @param {*} next 
    * @returns 
    */
-  delete: async (err, req, res, next) => {
+  delete: async (req, res, next) => {
     try {
+      if (!res.locals.note) return res.status(404).json({ message: "Note not found" })
+
       await res.locals.note.deleteOne()
       res.status(204).end()
     } catch (err) {
